Move App files into App folder.
