+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { store } from './app/store';
-import App from './App';
+import App from '../src/app/App'
 import reportWebVitals from './reportWebVitals';
 import './index.css';
